+++ ex_7/src/Gerente.cs
@@ -2,11 +2,11 @@ namespace ex_7.src
 {
     public class Gerente : Funcionario
     {
-        public Gerente(string nome, int idade, int salario) : base (nome, idade, salario)
+        public Gerente(string nome, int idade, double salario) : base (nome, idade, salario)
         {
 
         }
-        public override float Bonificacao()
+        public override double Bonificacao()
         {
             return this.salario + 10000;
         }
