+++ ex_7/src/Vendedor.cs
@@ -2,11 +2,11 @@ namespace ex_7.src
 {
     public class Vendedor : Funcionario
     {
-        public Vendedor(string nome, int idade, int salario) : base (nome, idade, salario)
+        public Vendedor(string nome, int idade, double salario) : base (nome, idade, salario)
         {
 
         }
-        public override float Bonificacao()
+        public override double Bonificacao()
         {
             return this.salario + 3000;
         }
